+++ pages/boards/[id]/components/NoCommentItem/styles.module.scss
