+++ pages/boards/components/AllArticle/AllArticleItem/index.tsx
@@ -1,5 +1,5 @@
 import Image from "next/image";
-import styles from "./styles.module.scss";
+import styles from "@/pages/boards/components/AllArticle/AllArticleItem/styles.module.scss";
 import heartIcon from "@/assets/icons/ic_heart.svg";
 import profileImage from "@/assets/icons/ic_profile.svg";
 import { Article } from "@/services/articles";
