+++ pages/boards/components/AllArticle/styles.module.scss
