+++ pages/boards/components/BestArticle/styles.module.scss
